utils/requirements.txt
elasticsearch==7.15.1
pandas
requests_aws4auth
PyYAML